+++ core/notif_nodemailer.js
@@ -101,7 +101,8 @@ module.exports = {
     },
 
     // todo: should be factorized with add, as there is only small difference
-    remove: async function(email) {
+    // eslint-disable-next-line no-unused-vars
+    remove: async function(email, sendmail=true) {
         if (CONFIG.nodemailer.list) {
             if(email===undefined ||email===null || email=='' || ! mail_set) {
                 return;